requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
pymupdf